API change: burstiness is not expressed in terms of messages
And not in terms of "burst rate", since it's not a term that can be
defined precisely, or something that is strictly guaranteed by the token
bucket algorithm. What it does guarantee is a maximum number of burst
messages, but it does not impose a time interval for them. Esentially,
those B messages go unthrottled, so the "burst rate" is always the
maximum throughput.