Add :granularity option (WIP: test, docs)
This option lets users choose how tightly the goal rate is enforced. By
default it is set to 1, which means every message has to potentially
wait for a token before being let through. This causes an almost
perfectly linear Messages vs. time traffic shape.

By choosing a larger granularity number, the number of messages that are
piped per token increases. The goal rate is still the same (the token
frequency arrival becomes lower to compensate), but the shape of the
Messages vs. time curve becomes more like a staircase. This can be
useful for cases where the goal is just to limit the number of messages
per unit of time, without caring for the traffic shape.