Make high granularities work with statistical mux
Instead of forwarding token-value messages from the bucket channel to
the user channel, we *insert* token-value tokens in the bucket every
sleep-time milliseconds. This makes it possible for each of these
token-value tokens to be taken by different channels, which may be the
case when either chan-throttler or fn-throttler are used.

Before, all token-value messages would have been available only to a
single channel (or function), possibly starving other channels (or
functions) sharing the same throttler.